+++ scripts/macros/spells/acidArrow.js
@@ -1,7 +1,6 @@
 import {chris} from '../../helperFunctions.js';
 import {queue} from '../../queue.js';
 async function attack(workflow) {
-    console.log('Here');
 	if (!workflow.isFumble) return;
     let queueSetup = await queue.setup(workflow.item.uuid, 'acidArrow', 50);
     if (!queueSetup) return;
@@ -11,7 +10,6 @@ async function attack(workflow) {
     queue.remove(workflow.item.uuid);
 }
 async function damage(workflow) {
-    console.log('Here too');
     if (workflow.hitTargets.size != 0) return;
     let queueSetup = await queue.setup(workflow.item.uuid, 'acidArrow', 50);
     if (!queueSetup) return;